backend-mvp/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-Cors
python-dotenv
flasgger
pytest
